Refactor test utils
setup/teardown code was detected as test functions.
